mywork/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
albumentations=1.0.3=pyhd8ed1ab_0
apex=0.1=pypi_0
astor=0.8.1=py37haa95532_0
atomicwrites=1.4.0=py_0
attrs=21.4.0=pyhd3eb1b0_0
blas=1.0=mkl
blosc=1.21.0=h19a0ad4_0
bottleneck=1.3.2=py37h2a96729_1
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py37h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2021.10.26=haa95532_4
cached-property=1.5.2=pypi_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py37haa95532_2
cffi=1.15.0=py37h2bbff1b_1
cfitsio=3.470=he774522_6
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
commonmark=0.9.1=pypi_0
configparser=5.0.2=pyhd3eb1b0_0
cryptography=36.0.0=py37h21b164f_0
cudatoolkit=10.2.89=h74a9793_1
cudnn=7.6.5=cuda10.2_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.25=py37h604cdb4_0
cytoolz=0.11.0=py37he774522_0
dask-core=2021.10.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
decorator=4.4.2=pyhd3eb1b0_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
easydict=1.9=pypi_0
faiss=1.7.1=py37cuda102h577d055_1_cuda
ffmpeg=4.2.2=he774522_0
flatbuffers=2.0=pypi_0
fonttools=4.28.5=pypi_0
freeglut=3.0.0=h6538335_5
freetype=2.10.4=hd328e21_0
fsspec=2022.1.0=pyhd3eb1b0_0
gast=0.3.3=py_0
geos=3.8.0=h33f27b4_0
giflib=5.2.1=h62dcd97_0
git=2.34.1=haa95532_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
google-auth=2.4.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.43.0=pypi_0
h5py=3.6.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.1=h6c2663c_0
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py37ha1f97ea_0
imageio=2.9.0=pyhd3eb1b0_0
imageio-ffmpeg=0.4.5=pyhd8ed1ab_0
imgaug=0.4.0=pyhd3eb1b0_0
importlib-metadata=4.8.2=py37haa95532_0
importlib_metadata=4.8.2=hd3eb1b0_0
imutils=0.5.4=py37h03978a9_1
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2022.0.0=haa95532_3663
jasper=2.0.14=hea7d32e_2
joblib=1.1.0=pypi_0
jpeg=9d=h2bbff1b_0
kiwisolver=1.3.2=pypi_0
lap=0.4.0=py37h9386db6_1004
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libblas=3.9.0=12_win64_mkl
libcblas=3.9.0=12_win64_mkl
libclang=11.1.0=default_h5c34c98_1
libdeflate=1.8=h2bbff1b_5
libfaiss=1.7.1=cuda102h4b21b0d_1_cuda
libfaiss-avx2=1.7.1=cuda102h1234567_1_cuda
libiconv=1.15=h1df5818_7
liblapack=3.9.0=12_win64_mkl
liblapacke=3.9.0=12_win64_mkl
libopencv=4.5.1=py37ha0199f4_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.19.1=h23ce68f_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.0=h2bbff1b_0
libwebp-base=1.2.0=h2bbff1b_0
libxml2=2.9.12=h0ad7f3c_0
libxslt=1.1.34=he774522_0
libzopfli=1.0.3=ha925a31_0
locket=0.2.1=py37haa95532_1
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.6=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-base=3.5.0=py37h6214cd6_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
more-itertools=8.12.0=pyhd3eb1b0_0
motmetrics=1.2.0=py_0
moviepy=1.0.1=py_0
munch=2.5.0=pypi_0
munkres=1.1.4=py_0
networkx=2.6.3=pyhd3eb1b0_0
numexpr=2.8.1=py37hb80d3ca_0
numpy=1.21.2=py37hfca59bb_0
numpy-base=1.21.2=py37h0829f74_0
nvidia-apex=0.1=py37h8e62851_4
oauthlib=3.1.1=pypi_0
olefile=0.46=py37_0
onnxruntime=1.10.0=pypi_0
opencv=4.5.1=py37h03978a9_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1m=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
paddlepaddle-gpu=2.2.1=py37_gpu_cuda10.2_windows
pandas=1.3.5=py37h6214cd6_0
partd=1.2.0=pyhd3eb1b0_0
pathtools=0.1.2=pyhd3eb1b0_1
pillow=9.0.0=pypi_0
pip=21.3.1=pypi_0
pluggy=0.13.1=py37haa95532_0
pretrainedmodels=0.7.4=pypi_0
proglog=0.1.9=py_0
promise=2.3=py37haa95532_0
protobuf=3.19.1=py37hd77b12b_0
psutil=5.8.0=py37h2bbff1b_1
py=1.10.0=pyhd3eb1b0_0
py-opencv=4.5.1=py37heaed05f_0
pyaml=21.10.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycocotools-windows=2.0.0.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pypi_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.6=pypi_0
pyside2=5.13.2=py37h760f651_4
pysocks=1.7.1=py37_1
pytest=6.2.4=py37haa95532_2
python=3.7.11=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.10.1=py3.7_cuda10.2_cudnn7_0
pytorch-metric-learning=1.1.0=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py37he774522_2
pyyaml=6.0=pypi_0
qt=5.12.9=h5909a2a_4
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rich=11.0.0=pypi_0
rsa=4.8=pypi_0
scikit-image=0.18.3=py37hf11a4ad_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py37h0a974cb_0
seaborn=0.11.2=pypi_0
sentry-sdk=1.5.4=pyhd8ed1ab_0
setuptools=60.5.0=pypi_0
shapely=1.7.1=py37h06580b3_0
shortuuid=1.0.8=py37h03978a9_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.8=h33f27b4_0
sqlite=3.37.0=h2bbff1b_0
subprocess32=3.5.4=py_1
tabulate=0.8.9=pypi_0
tbb=2021.5.0=h59b6b97_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
thop=0.0.31-2005241907=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2021.7.2=pyhd3eb1b0_2
timm=0.4.12=pypi_0
tk=8.6.11=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.10.1=py37_cu102
torchvision=0.11.2=pypi_0
tqdm=4.62.3=pyhd3eb1b0_1
typing-extensions=4.0.1=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.29.30037=h902a5da_6
wandb=0.12.10=pyhd8ed1ab_0
werkzeug=2.0.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
xmltodict=0.12.0=pyhd3eb1b0_0
xz=5.2.5=h62dcd97_0
yacs=0.1.8=pypi_0
yaml=0.2.5=he774522_0
yaspin=2.1.0=pyhd8ed1ab_0
zfp=0.5.5=hd77b12b_6
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h8cc25b3_4
zstd=1.4.9=h19a0ad4_0
